Power rates are now doubles to support accuracy in the sub-mW range.

+++ trunk/org.qualitune.jouleunit/src/org/qualitune/jouleunit/AbstractEnergyProfile.java
@@ -88,7 +88,7 @@ public abstract class AbstractEnergyProfile implements EnergyProfile {
 	 * 
 	 * @see org.qualitune.jouleunit.EnergyProfile#getPeakPowerRate()
 	 */
-	public long getPeakPowerRate() {
+	public double getPeakPowerRate() {
 
 		return getPeakPowerRate(START_EVENT_ID, END_EVENT_ID);
 	}
